+++ project_composer/cli/purge.py
@@ -20,8 +20,8 @@ from .base_options import COMMON_OPTIONS
     **COMMON_OPTIONS["repository"]["kwargs"]
 )
 @click.option(
-    *COMMON_OPTIONS["syspath"]["args"],
-    **COMMON_OPTIONS["syspath"]["kwargs"]
+    *COMMON_OPTIONS["syspaths"]["args"],
+    **COMMON_OPTIONS["syspaths"]["kwargs"]
 )
 @click.option(
     "--commit",
@@ -31,7 +31,7 @@ from .base_options import COMMON_OPTIONS
     ),
 )
 @click.pass_context
-def purge_command(context, manifest, repository, syspath, commit):
+def purge_command(*args, **parameters):
     """
     Purge applications repository from any directory that is not an enabled application
     module.
@@ -43,26 +43,30 @@ def purge_command(context, manifest, repository, syspath, commit):
     """
     logger = logging.getLogger(__pkgname__)
 
-    logger.debug("Using manifest: {}".format(manifest))
-    manifest = Manifest.load(manifest)
+    logger.debug("Using manifest: {}".format(parameters["manifest"]))
+    manifest = Manifest.load(parameters["manifest"])
 
-    # Override some manifest options from arguments
-    if repository is not None:
-        manifest.repository = repository
-    if syspath is not None:
-        manifest.syspaths.extend(syspath)
+    # Override base manifest settings from given arguments
+    for name in manifest.get_fields():
+        if name != "requirements" and parameters.get(name) is not None:
+            setattr(manifest, name, parameters.get(name))
 
-    if manifest.repository:
-        logger.debug("Applications repository: {}".format(manifest.repository))
-    else:
-        logger.critical("Applications repository is required.")
+    if not manifest.repository:
+        logger.critical(
+            "Applications repository is required for this command, read help for "
+            "more details."
+        )
         raise click.Abort()
+
+    # Logging used settings
+    logger.debug("Applications repository: {}".format(manifest.repository))
 
     for item in manifest.syspaths:
         logger.debug("Loading in sys.path: {}".format(item))
 
     composer = PurgeApplications(manifest)
 
+    commit = parameters.get("commit")
     if commit:
         to_remove = composer.commit()
     else: